arch: Move *_getsp to the common place arch/arch.h
so other place can get the stack pointer easily
