+++ arch/hc/include/arch.h
@@ -55,6 +55,21 @@
  ****************************************************************************/
 
 /****************************************************************************
+ * Name: hc_getsp
+ ****************************************************************************/
+
+static inline uint16_t hc_getsp(void)
+{
+  uint16_t ret;
+  __asm__
+  (
+    "\tsts %0\n"
+  : "=m"(ret) :
+  );
+  return ret;
+}
+
+/****************************************************************************
  * Public Types
  ****************************************************************************/
