+++ arch/hc/include/tls.h
@@ -35,21 +35,6 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: hc_getsp
- ****************************************************************************/
-
-static inline uint16_t hc_getsp(void)
-{
-  uint16_t ret;
-  __asm__
-  (
-    "\tsts %0\n"
-  : "=m"(ret) :
-  );
-  return ret;
-}
-
-/****************************************************************************
  * Name: up_tls_info
  *
  * Description:
