+++ arch/mips/include/arch.h
@@ -55,6 +55,21 @@
  ****************************************************************************/
 
 /****************************************************************************
+ * Name: mips_getsp
+ ****************************************************************************/
+
+static inline uint32_t mips_getsp(void)
+{
+  register uint32_t sp;
+  __asm__
+  (
+    "\tadd  %0, $0, $29\n"
+    : "=r"(sp)
+  );
+  return sp;
+}
+
+/****************************************************************************
  * Public Types
  ****************************************************************************/
