+++ arch/mips/include/tls.h
@@ -35,21 +35,6 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: mips_getsp
- ****************************************************************************/
-
-static inline uint32_t mips_getsp(void)
-{
-  register uint32_t sp;
-  __asm__
-  (
-    "\tadd  %0, $0, $29\n"
-    : "=r"(sp)
-  );
-  return sp;
-}
-
-/****************************************************************************
  * Name: up_tls_info
  *
  * Description:
