+++ arch/misoc/include/tls.h
@@ -35,19 +35,6 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: misoc_getsp
- ****************************************************************************/
-
-static inline uint32_t misoc_getsp(void)
-{
-  register uint32_t sp;
-
-  __asm__ __volatile__("addi %0, sp, 0" : "=r" (sp));
-
-  return sp;
-}
-
-/****************************************************************************
  * Name: up_tls_info
  *
  * Description:
