+++ arch/or1k/include/arch.h
@@ -88,6 +88,23 @@
  ****************************************************************************/
 
 /****************************************************************************
+ * Name: or1k_getsp
+ ****************************************************************************/
+
+static inline uint32_t or1k_getsp(void)
+{
+  uint32_t sp;
+
+  __asm__
+  (
+    "\tmov %0, sp\n\t"
+    : "=r"(sp)
+  );
+
+  return sp;
+}
+
+/****************************************************************************
  * Public Types
  ****************************************************************************/
 
@@ -98,8 +115,8 @@
  * nuttx/arch/<architecture>/include/arch.h.
  *
  * These tables would hold the physical address of the level 2 page tables.
- * All would be initially NULL and would not be backed up with physical memory
- * until mappings in the level 2 page table are required.
+ * All would be initially NULL and would not be backed up with physical
+ * memory until mappings in the level 2 page table are required.
  */
 
 struct group_addrenv_s
