+++ arch/or1k/include/tls.h
@@ -35,23 +35,6 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: or1k_getsp
- ****************************************************************************/
-
-static inline uint32_t or1k_getsp(void)
-{
-  uint32_t sp;
-
-  __asm__
-  (
-    "\tmov %0, sp\n\t"
-    : "=r"(sp)
-  );
-
-  return sp;
-}
-
-/****************************************************************************
  * Name: up_tls_info
  *
  * Description:
